Create README.md
This project is the front-end part of a form to register candidates for vacancies in the programming area. The candidate must fill in all the fields according to the expected format, if any of these fields are not filled in correctly, the system issues error messages informing exactly the problems and in which fields they occur. In addition to text entry fields, there is a field for the candidate to upload a PDF file corresponding to the candidate's CV.

The main purpose of this project is the study and application of the react-form-hook libraries, which offer several functionalities for form validation and the Tailwind CSS. which offers several ready-to-use CSS classes. In addition to these, react-international-phone libraries were used to validate the telephone number field and react-tailwindcss-select to create a select box with available job positions.